csv_service/requirements.txt
django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
python-decouple==3.3
